load user info from index page, and fix timing issues, reducing the number of times calling the persons endppoint
